+++ tool/utils.ts
@@ -38,6 +38,11 @@ const ARCH: 'ia32' | 'x64' = (() => {
       return 'ia32';
     case 'x64':
       return 'x64';
+    // TODO: This is blocked until Github Actions supports compiling Dart Sass
+    // for arm64. Until then, download the x64 binary for arm64 users.
+    // https://github.com/sass/dart-sass/issues/1125
+    case 'arm64':
+      return 'x64';
     default:
       throw Error(`Architecure ${process.arch} is not supported.`);
   }